Fuzz cleanup
