+++ src/agent0/hyperfuzz/__init__.py
@@ -1,0 +1,1 @@
+from .fuzz_assertion_exception import FuzzAssertionException
