+++ src/agent0/hyperfuzz/unit_fuzz/fuzz_long_short_maturity_values.py
@@ -30,13 +30,10 @@ from fixedpointmath import FixedPoint, isclose
 from agent0.core.hyperdrive.crash_report import build_crash_trade_result, log_hyperdrive_crash_report
 from agent0.core.hyperdrive.interactive import ILocalChain, ILocalHyperdrive
 from agent0.core.hyperdrive.interactive.event_types import CloseLong, CloseShort, OpenLong, OpenShort
-from agent0.core.interactive_fuzz.helpers import (
-    FuzzAssertionException,
-    advance_time_after_checkpoint,
-    execute_random_trades,
-    setup_fuzz,
-)
+from agent0.hyperfuzz import FuzzAssertionException
 from agent0.hypertypes.fixedpoint_types import CheckpointFP
+
+from .helpers import advance_time_after_checkpoint, execute_random_trades, setup_fuzz
 
 # main script has a lot of stuff going on
 # pylint: disable=too-many-locals
