+++ src/agent0/hyperfuzz/unit_fuzz/fuzz_present_value.py
@@ -30,7 +30,9 @@ from fixedpointmath import FixedPoint, isclose
 from agent0.core.hyperdrive import HyperdriveActionType
 from agent0.core.hyperdrive.crash_report import build_crash_trade_result, log_hyperdrive_crash_report
 from agent0.core.hyperdrive.interactive import ILocalChain, ILocalHyperdrive
-from agent0.core.interactive_fuzz.helpers import FuzzAssertionException, setup_fuzz
+from agent0.hyperfuzz import FuzzAssertionException
+
+from .helpers import setup_fuzz
 
 # tests have lots of stuff
 # pylint: disable=too-many-locals
