+++ src/agent0/hyperfuzz/unit_fuzz/helpers/__init__.py
@@ -3,5 +3,4 @@
 from .advance_time import advance_time_after_checkpoint, advance_time_before_checkpoint
 from .close_random_trades import close_trades, permute_trade_events
 from .execute_random_trades import execute_random_trades
-from .fuzz_assertion_exception import FuzzAssertionException
 from .setup_fuzz import setup_fuzz
