+++ src/agent0/hyperfuzz/unit_fuzz/helpers/close_random_trades.py
